Init commit: working config for compiling with embassy
Push config w/ which we can build/flash and
monitor a `no-std` binary running embassy on the
esp32c3 with a simple `cargo run --release`.

The binary needs to be build with nightly in order
to use `type_alias_impl_trait` feature which in turn
allows embassy to statically allocate all tasks, instead
of allocating them from an arena.

Signed-off-by: Gasper Stukelj <gasperstukelj@proton.me>