Enable temperature sensing
Enable the esp32-c3 built-in temperature sensor.

The readings are printed raw (without computing
the actual temperature). There's no HAL impl, so
we're using board PAC crate (re-exported by HAL)
to set up the sensor. We are "stealing" two
singleton structs which is unsafe due to possible
concurrent mutable access, but we're only setting
register bits pertaining to the temperature
sensor, so this is justified.

Signed-off-by: Gasper Stukelj <gasperstukelj@proton.me>