Button: init IO driver
Constructing the IO struct enables and configures
board IO driver as a side effect. We don't need
the struct and can mark the object as unused, but
the compiler is not allowed to optimize away the
side effects, however, we can place high confidence
in it optimizing away struct construction/allocation.