Netstack: Connect socket to public broker
TODO: Find out what's going on with the DHCP
config. As far as I can tell the DHCP server
pushes a DNS server address as well, but it
get's lost somewhere in the embassy-net stack
and we can run into a panic.

Temporarily bypassed this issue by hardcoding
1.1.1.1 as the DNS server regardless of the
config passed. The following was changed in the
embassy_net/src/lib.rs cached files:

```
  let addr = IpAddress::Ipv4(Ipv4Address::new(1, 1, 1, 1));
  // Apply DNS servers
  #[cfg(feature = "dns")]
  s.sockets
      .get_mut::<smoltcp::socket::dns::Socket>(self.dns_socket)
      .update_servers(&[addr]);
```